+++ src/components/profile/ProfileForm.tsx
@@ -2,7 +2,7 @@
 
 import { createMoverProfile } from "@/actions/profile/create-moverProfile.action";
 import SolidButton from "@/components/common/buttons/SolidButton";
-import InputField from "@/components/profile/InputField";
+import GeneralInputField from "@/components/profile/GeneralInputField";
 import React, { useActionState, useState } from "react";
 import {
   validateName,
@@ -11,11 +11,15 @@ import {
   validateDetailDescription,
   validateServiceType,
   validateArea,
-} from "@/validations";
+} from "@/validations"; //주석: 입력값의 유효성 검사를 개별적으로 진행
+import ImageInputField from "./ImageInputField";
+import ButtonInputField from "./ButtonInputField";
+import TextAreaInputField from "./TextAreaInputField";
 
 function ProfileForm() {
   const [, formAction, isPending] = useActionState(createMoverProfile, null);
 
+  //주석: 시작하기 버튼 활성화를 위한 상태 관리
   const [fieldValidity, setFieldValidity] = useState<Record<string, boolean>>({
     name: false,
     career: false,
@@ -25,7 +29,7 @@ function ProfileForm() {
     area: false,
   });
 
-  // InputField에서 호출할 함수
+  //주석: 시작하기 버튼 활성화를 위해 InputField 컴포넌트로 내려줄 함수
   const handleValidityChange = (name: string, isValid: boolean) => {
     setFieldValidity((prev) => ({
       ...prev,
@@ -33,6 +37,7 @@ function ProfileForm() {
     }));
   };
 
+  //주석: 시작하기 버튼 활성화
   const isDisabled =
     isPending || !Object.values(fieldValidity).every((v) => v === true);
 
@@ -42,12 +47,12 @@ function ProfileForm() {
       className="flex flex-col w-full mt-6 lg:mt-12 lg:flex-row lg:gap-18"
     >
       <div className="flex flex-col flex-1">
-        <InputField name="image" isImage={true} text="프로필 이미지" />
+        <ImageInputField text="프로필 이미지" />
 
         <hr className="hidden lg:block m-0 p-0 border-t-[1px] border-line-100 my-8" />
 
         <div className="mt-8">
-          <InputField
+          <GeneralInputField
             name="name"
             text="별명"
             placeholder="사이트에 노출될 이름을 입력해주세요"
@@ -61,58 +66,57 @@ function ProfileForm() {
 
         <hr className="m-0 p-0 border-t-[1px] border-line-100 my-8" />
 
-        <InputField
+        <GeneralInputField
           name="career"
           text="경력"
           placeholder="기사님의 경력을 입력해주세요"
           height="h-13 mg:h-13 lg:h-16"
           validator={validateCareer}
-          onValidChange={(name, isValid) => handleValidityChange(name, isValid)}
+          onValidChange={handleValidityChange}
         />
 
         <hr className="m-0 p-0 border-t-[1px] border-line-100 my-8" />
 
-        <InputField
+        <GeneralInputField
           name="onelineIntroduction"
           text="한 줄 소개"
           placeholder="한 줄 소개를 입력해주세요"
           height="h-13 mg:h-13 lg:h-16"
           validator={validateOnelineIntroduction}
-          onValidChange={(name, isValid) => handleValidityChange(name, isValid)}
+          onValidChange={handleValidityChange}
         />
 
         <hr className="m-0 p-0 border-t-[1px] border-line-100 my-8 lg:hidden" />
       </div>
 
       <div className="flex-1">
-        <InputField
+        <TextAreaInputField
           name="detailDescription"
           text="상세 설명"
-          isTextArea={true}
           placeholder="상세 내용을 입력해주세요"
           height="h-[160px]"
           validator={validateDetailDescription}
-          onValidChange={(name, isValid) => handleValidityChange(name, isValid)}
+          onValidChange={handleValidityChange}
         />
 
         <hr className="m-0 p-0 border-t-[1px] border-line-100 my-8" />
 
-        <InputField
+        <ButtonInputField
           name="serviceType"
           text="제공 서비스"
           isServiceType={true}
           validator={validateServiceType}
-          onValidChange={(name, isValid) => handleValidityChange(name, isValid)}
+          onValidChange={handleValidityChange}
         />
 
         <hr className="m-0 p-0 border-t-[1px] border-line-100 my-8" />
 
-        <InputField
+        <ButtonInputField
           name="area"
           text="서비스 가능 지역"
           isArea={true}
           validator={validateArea}
-          onValidChange={(name, isValid) => handleValidityChange(name, isValid)}
+          onValidChange={handleValidityChange}
         />
 
         <div className="mt-17">
